Fix bug in routing
