+++ Pipfile
@@ -11,7 +11,12 @@ tensorflow = "==2.*"
 tensorflow-datasets = "*"
 absl-py = ">=0.7.0"
 numpy = "*"
+#more_itertools = ">=8.3"
 importlib-resources = {version = "*", markers="python_version < '3.7'"}
 
 # Optional packages
 # pyyaml = "*"        # for serialization and deserialization of models from YAML
+
+# This conflict was resolved by taking the more restrictive of the two versions
+# more_itertools = {version = ">=8.3", markers="python_version < '3.10'"}
+# more_itertools = ">=4.1"
