+++ yolo/utils/scripts/darknet2tf/_read_weights.py
@@ -7,7 +7,6 @@ from __future__ import annotations
 import io
 import numpy as np
 import os
-import itertools
 
 from typing import Union
 
@@ -15,20 +14,33 @@ from .config_classes import *
 from ..dn2dicts import convertConfigFile
 from ...file_manager import PathABC, get_size, open_if_not_open
 
+import more_itertools
+
+
+def _shift(n):
+    """Shifts a Python index 1 to the left to ignore [net] section"""
+    if n < 0:
+        return n
+    return n + 1
 
 def build_layer(layer_dict, file, prevlayer, net):
     """consturct layer and load weights from file"""
+    net = more_itertools.SequenceView(net)
+
     layer = layer_builder[layer_dict['_type']].from_dict(prevlayer, layer_dict)
 
     # temprary use for management of routing layers
     if layer._type == "route":
         layers = layer_dict['layers']
+        print(f"{layer} {layers}")
         if type(layers) is tuple:
-            route = layers[1]
-            prevlayer = net[route + 1]
-            layer.c += prevlayer.shape[-1]
+            layer.c = 0
+            for l in layers:
+                lshape = net[_shift(l)].shape
+                layer.c += lshape[-1]
         else:
-            layer.c //= 2
+            lshape = net[_shift(layers)].shape
+            layer.c = lshape[-1]
 
     if file is not None:
         bytes_read = layer.load_weights(file)
@@ -66,7 +78,7 @@ def read_file(config, weights):
     outputs = [None]
     for layer_dict in config:
         if layer_dict["_type"] != "decoder_encoder_split":
-            layer, num_read = build_layer(layer_dict, weights, net[-1], encoder)
+            layer, num_read = build_layer(layer_dict, weights, net[-1], encoder + decoder)
             if layer_dict["_type"] != 'yolo' and layer.shape[-1] != 255:
                 net.append(layer)
             else:
@@ -80,48 +92,6 @@ def read_file(config, weights):
     del encoder[0]
     del decoder[0]
     return encoder, decoder, outputs, bytes_read
-
-
-def interleve_weights(block):
-    """merge weights to fit the DarkResnet block style"""
-    if len(block) == 0:
-        return []
-    weights_temp = []
-    for layer in block:
-        weights = layer.get_weights()
-        weights = [tuple(weights[0:3]), tuple(weights[3:])]
-        weights_temp.append(weights)
-    top, bottom = tuple(zip(*weights_temp))
-    weights = list(itertools.chain.from_iterable(top)) + \
-        list(itertools.chain.from_iterable(bottom))
-    return weights
-
-
-def get_darknet53_tf_format(net, only_weights=True):
-    """convert weights from darknet sequntial to tensorflow weave, Darknet53 Backbone"""
-    combo_blocks = []
-    for i in range(2):
-        layer = net.pop(0)
-        combo_blocks.append(layer)
-    # ugly code i will document, very tired
-    encoder = []
-    while len(net) != 0:
-        blocks = []
-        layer = net.pop(0)
-        while layer._type != "shortcut":
-            blocks.append(layer)
-            layer = net.pop(0)
-        encoder.append(blocks)
-    new_net = combo_blocks + encoder
-    weights = []
-    if only_weights:
-        for block in new_net:
-            if type(block) != list:
-                weights.append(block.get_weights())
-            else:
-                weights.append(interleve_weights(block))
-    print("converted/interleved weights for tensorflow format")
-    return new_net, weights
 
 
 def read_weights(config_file: Union[PathABC, io.TextIOBase],
